requirements.txt
Django==4.0.6
django-phonenumber-field==6.3.0